fix: rename useSortableTable hook
